Rename files in impulse-form
