+++ packages/react-impulse-form/src/impulse-form-shape.ts
@@ -1,24 +1,23 @@
 import type Types from "ts-toolbelt"
 
+import { isTrue } from "~/is-true"
+import { isBoolean } from "~/is-boolean"
+import { isTruthy } from "~/is-truthy"
+import { isString } from "~/is-string"
+import { isNull } from "~/is-null"
+import { isUndefined } from "~/is-undefined"
+import { isFunction } from "~/is-function"
+import type { Setter } from "~/setter"
+import type { Compute } from "~/compute"
+import { params } from "~/params"
+
 import { type Scope, batch } from "./dependencies"
-import {
-  type ComputeObject,
-  isTrue,
-  type Setter,
-  params,
-  isUndefined,
-  isString,
-  isBoolean,
-  isFunction,
-  isTruthy,
-  isNull,
-} from "./utils"
 import {
   type GetImpulseFormParam,
   type ImpulseFormParamsKeys,
   ImpulseForm,
-} from "./ImpulseForm"
-import { VALIDATE_ON_TOUCH, type ValidateStrategy } from "./ValidateStrategy"
+} from "./impulse-form"
+import { VALIDATE_ON_TOUCH, type ValidateStrategy } from "./validate-strategy"
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type ImpulseFormShapeFields = Record<string | number, any>
@@ -27,7 +26,7 @@ type ImpulseFormShapeParam<
   TFields extends ImpulseFormShapeFields,
   TKey extends ImpulseFormParamsKeys,
   TFallback extends "field" | "nothing" = "nothing",
-> = ComputeObject<
+> = Compute<
   Types.Object.Filter<
     {
       readonly [TField in Types.Any.Keys<TFields>]: GetImpulseFormParam<
