+++ packages/react-impulse-form/src/index.ts
@@ -1,7 +1,8 @@
-export * from "./ImpulseForm"
-export * from "./ImpulseFormValue"
-export * from "./ImpulseFormShape"
-export * from "./ImpulseFormList"
-export * from "./ZodLikeSchema"
-export type { ValidateStrategy } from "./ValidateStrategy"
-export type { Setter, Result } from "./utils"
+export * from "./impulse-form"
+export * from "./impulse-form-value"
+export * from "./impulse-form-shape"
+export * from "./impulse-form-list"
+export * from "./zod-like-schema"
+export type { ValidateStrategy } from "./validate-strategy"
+export type { Setter } from "~/setter"
+export type { Result } from "./result"
