+++ packages/react-impulse-form/src/result.ts
@@ -1,0 +1,3 @@
+export type Result<TError, TData> = [TError] extends [null]
+  ? [null, TData]
+  : [TError, null] | [null, TData]
