+++ packages/react-impulse-form/tests/ImpulseFormValue/setValidateOn.spec.ts
@@ -1,7 +1,8 @@
 import { z } from "zod"
 
+import type { Setter } from "~/setter"
+
 import {
-  type Setter,
   type ValidateStrategy,
   ImpulseFormValue,
   type ImpulseFormValueSchemaOptions,
