--- packages/react-impulse/src/is-function.ts
@@ -1,5 +1,0 @@
-export function isFunction<
-  TFunction extends (...args: ReadonlyArray<never>) => unknown,
->(anything: unknown): anything is TFunction {
-  return typeof anything === "function"
-}
