+++ packages/react-impulse/tsconfig.json
@@ -1,4 +1,11 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "tests", "./*.ts"]
+  "include": [
+    "src",
+    "tests",
+    "./*.ts",
+    "../../tools/has-property.ts",
+    "../../tools/is-strict-equal.ts",
+    "../../tools/identity.ts"
+  ]
 }
