+++ tools/compute.ts
@@ -1,0 +1,3 @@
+export type Compute<Obj> = unknown & {
+  [K in keyof Obj]: Obj[K]
+}
