+++ tools/has-property.ts
