+++ tools/is-array.ts
@@ -1,0 +1,3 @@
+export const isArray: <T>(
+  data: ArrayLike<unknown> | T,
+) => data is ReadonlyArray<unknown> = Array.isArray
