+++ tools/is-boolean.ts
@@ -1,0 +1,3 @@
+export function isBoolean(value: unknown): value is boolean {
+  return typeof value === "boolean"
+}
