+++ tools/is-defined.ts
@@ -1,0 +1,5 @@
+export function isDefined<TValue>(
+  data: void | null | undefined | TValue,
+): data is TValue {
+  return data != null
+}
