+++ tools/is-false.ts
@@ -1,0 +1,3 @@
+export function isFalse(value: unknown): value is false {
+  return value === false
+}
