+++ tools/is-null.ts
@@ -1,0 +1,3 @@
+export function isNull(value: unknown): value is null {
+  return value === null
+}
