+++ tools/is-object.ts
@@ -1,0 +1,5 @@
+import { isDefined } from "~/is-defined"
+
+export function isObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && isDefined(value)
+}
