+++ tools/is-strict-equal.ts
