+++ tools/is-string.ts
@@ -1,0 +1,3 @@
+export function isString(value: unknown): value is string {
+  return typeof value === "string"
+}
