+++ tools/is-true.ts
@@ -1,0 +1,3 @@
+export function isTrue(value: unknown): value is true {
+  return value === true
+}
