+++ tools/is-undefined.ts
@@ -1,0 +1,3 @@
+export function isUndefined(value: unknown): value is undefined {
+  return value === undefined
+}
