+++ tools/null-or-non-nullable.ts
@@ -1,0 +1,3 @@
+export type NullOrNonNullable<T> = [T, null] extends [null, T]
+  ? null
+  : NonNullable<T>
